+++ qunit/tests_inputeventonly.js
@@ -5,7 +5,6 @@ define([
 	"../dist/inputmask/inputmask.extensions",
 	"../dist/inputmask/inputmask.numeric.extensions",
 	"../dist/inputmask/inputmask.phone.extensions",
-	"../dist/inputmask/inputmask.regex.extensions",
 	"prototypeExtensions",
 	"simulator"
 ], function (qunit, $, Inputmask) {