+++ qunit/tests_regex.js
@@ -5,18 +5,17 @@ define([
 	"../dist/inputmask/inputmask.extensions",
 	"../dist/inputmask/inputmask.numeric.extensions",
 	"../dist/inputmask/inputmask.phone.extensions",
-	"../dist/inputmask/inputmask.regex.extensions",
 	"prototypeExtensions",
 	"simulator"
 ], function(qunit, $, Inputmask) {
 
 qunit.module("Regex masks");
 
-qunit.test("inputmask(\"Regex\", { regex: \"[0-9]*\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[0-9]*\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[0-9]*"
 	}).mask(testmask);
 
@@ -26,11 +25,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[0-9]*\"});", function(assert) {
 	assert.equal(testmask.value, "12345", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[0-9]*\"}); ~ isComplete", function(assert) {
+qunit.test("inputmask({ regex: \"[0-9]*\"}); ~ isComplete", function(assert) {
 	var $fixture = $("#qunit-fixture"), done = assert.async();
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[0-9]*",
 		oncomplete: function() {
 			assert.equal(testmask.value, "1", "Result " + testmask.value);
@@ -42,12 +41,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[0-9]*\"}); ~ isComplete", function(
 	$("#testmask").SendKey("1");
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[A-Za-z\u0410-\u044F\u0401\u04510-9]*\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[A-Za-z\u0410-\u044F\u0401\u04510-9]*\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
-		regex: "[A-Za-z\u0410-\u044F\u0401\u04510-9]*"
+	Inputmask({regex: "[A-Za-z\u0410-\u044F\u0401\u04510-9]*"
 	}).mask(testmask);
 
 	testmask.focus();
@@ -56,11 +54,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[A-Za-z\u0410-\u044F\u0401\u04510-9]
 	assert.equal(testmask.value, "123abc45", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[A-Za-z\u0410-\u044F\u0401\u0451]+[A-Za-z\u0410-\u044F\u0401\u04510-9]*\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[A-Za-z\u0410-\u044F\u0401\u0451]+[A-Za-z\u0410-\u044F\u0401\u04510-9]*\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[A-Za-z\u0410-\u044F\u0401\u0451]+[A-Za-z\u0410-\u044F\u0401\u04510-9]*"
 	}).mask(testmask);
 
@@ -71,11 +69,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[A-Za-z\u0410-\u044F\u0401\u0451]+[A
 });
 
 
-qunit.test("inputmask(\"Regex\", { regex: \"[A-Za-z\u0410-\u044F\u0401\u0451]{1}[A-Za-z\u0410-\u044F\u0401\u04510-9]*\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[A-Za-z\u0410-\u044F\u0401\u0451]{1}[A-Za-z\u0410-\u044F\u0401\u04510-9]*\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[A-Za-z\u0410-\u044F\u0401\u0451]{1}[A-Za-z\u0410-\u044F\u0401\u04510-9]*"
 	}).mask(testmask);
 
@@ -85,11 +83,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[A-Za-z\u0410-\u044F\u0401\u0451]{1}
 	assert.equal(testmask.value, "abc45", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[-]?(([1-8][0-9])|[1-9]0?)"
 	}).mask(testmask);
 
@@ -99,11 +97,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", fu
 	assert.equal(testmask.value, "90", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[-]?(([1-8][0-9])|[1-9]0?)"
 	}).mask(testmask);
 
@@ -113,11 +111,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", fu
 	assert.equal(testmask.value, "", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", function(assert) {
+qunit.test("inputmask({ regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[-]?(([1-8][0-9])|[1-9]0?)"
 	}).mask(testmask);
 
@@ -127,11 +125,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[-]?(([1-8][0-9])|[1-9]0?)\"});", fu
 	assert.equal(testmask.value, "-78", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\" - simple regex email", function(assert) {
+qunit.test("inputmask({ regex: \"[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\" - simple regex email", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?"
 	}).mask(testmask);
 
@@ -141,11 +139,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?
 	assert.equal(testmask.value, "some.body@mail.com", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\" - complexer regex email", function(assert) {
+qunit.test("inputmask({ regex: \"[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\" - complexer regex email", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?\\.)+[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?"
 	}).mask(testmask);
 
@@ -155,11 +153,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"[a-za-zA-Z0-9!#$%&'*+/=?^_`{|}~-]+(?
 	assert.equal(testmask.value, "denise.van.de.cruys@mail.com", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 1-123-4562", function(assert) {
+qunit.test("inputmask({ regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 1-123-4562", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))"
 	}).mask(testmask);
 
@@ -168,11 +166,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|(
 
 	assert.equal(testmask.value, "1-123-4562", "Result " + testmask.value);
 });
-qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 20-222-2222", function(assert) {
+qunit.test("inputmask({ regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 20-222-2222", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))"
 	}).mask(testmask);
 
@@ -181,11 +179,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|(
 
 	assert.equal(testmask.value, "20-222-2222", "Result " + testmask.value);
 });
-qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 22-222-234", function(assert) {
+qunit.test("inputmask({ regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 22-222-234", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))"
 	}).mask(testmask);
 
@@ -195,11 +193,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|(
 	assert.equal(testmask.value, "22-222-234", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 70-12-34", function(assert) {
+qunit.test("inputmask({ regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))\" - mrpanacs regex 70-12-34", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|((1|30|20|70)-([0-9]{3}-[0-9]{4}))"
 	}).mask(testmask);
 
@@ -209,11 +207,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"(([2-9][0-9])-([0-9]{3}-[0-9]{3}))|(
 	assert.equal(testmask.value, "70-123", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?\" - arame regex 12", function(assert) {
+qunit.test("inputmask({ regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?\" - arame regex 12", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?"
 	}).mask(testmask);
 
@@ -223,11 +221,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|
 	assert.equal(testmask.value, "12", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?\" } - arame regex 12.5", function(assert) {
+qunit.test("inputmask({ regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?\" } - arame regex 12.5", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?"
 	}).mask(testmask);
 
@@ -237,11 +235,11 @@ qunit.test("inputmask(\"Regex\", { regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|
 	assert.equal(testmask.value, "12.5", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", { regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?\" } - arame regex 12.75", function(assert) {
+qunit.test("inputmask({ regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?\" } - arame regex 12.75", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|75))?"
 	}).mask(testmask);
 
@@ -251,12 +249,13 @@ qunit.test("inputmask(\"Regex\", { regex: \"([0-9]|[1][0-9]|[2][0-3]?)(\\.(5|25|
 	assert.equal(testmask.value, "12.75", "Result " + testmask.value);
 });
 
-qunit.test("inputmask('Regex', { regex: \"(abc)+(def)\" }); - Flyarbonkers regex abcdef", function(assert) {
+qunit.test("inputmask({ regex: \"(abc)+(def)\" }); - Flyarbonkers regex abcdef", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask('Regex', {
-		regex: "(abc)+(def)"
+	Inputmask({
+		regex: "(abc)+(def)",
+		jitMasking: true
 	}).mask(testmask);
 
 	testmask.focus();
@@ -265,12 +264,13 @@ qunit.test("inputmask('Regex', { regex: \"(abc)+(def)\" }); - Flyarbonkers regex
 	assert.equal(testmask.value, "abcdef", "Result " + testmask.value);
 });
 
-qunit.test("inputmask('Regex', { regex: \"(abc)+(def)\" }); - Flyarbonkers regex 123a4b5c6d7e8f", function(assert) {
+qunit.test("inputmask({ regex: \"(abc)+(def)\" }); - Flyarbonkers regex 123a4b5c6d7e8f", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask('Regex', {
-		regex: "(abc)+(def)"
+	Inputmask({
+		regex: "(abc)+(def)",
+		jitMasking: true
 	}).mask(testmask);
 
 	testmask.focus();
@@ -279,12 +279,13 @@ qunit.test("inputmask('Regex', { regex: \"(abc)+(def)\" }); - Flyarbonkers regex
 	assert.equal(testmask.value, "abcdef", "Result " + testmask.value);
 });
 
-qunit.test("inputmask('Regex', { regex: \"(abc)+(def)\" }); - Flyarbonkers regex abcabcdef", function(assert) {
+qunit.test("inputmask({ regex: \"(abc)+(def)\" }); - Flyarbonkers regex abcabcdef", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask('Regex', {
-		regex: "(abc)+(def)"
+	Inputmask({
+		regex: "(abc)+(def)",
+		jitMasking: true
 	}).mask(testmask);
 
 	testmask.focus();
@@ -293,25 +294,26 @@ qunit.test("inputmask('Regex', { regex: \"(abc)+(def)\" }); - Flyarbonkers regex
 	assert.equal(testmask.value, "abcabcdef", "Result " + testmask.value);
 });
 
-qunit.test("inputmask('Regex', { regex: \"(abc){2,4}(def)\" }); - Flyarbonkers regex abdefcafebcaefbfcabcdef", function(assert) {
+qunit.test("inputmask({ regex: \"(abc){2,4}(def)\" }); - Flyarbonkers regex abcafebcaefbfcabcdef", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask('Regex', {
-		regex: "(abc){2,4}(def)"
+	Inputmask({
+		regex: "(abc){2,4}(def)",
+		jitMasking: true
 	}).mask(testmask);
 
 	testmask.focus();
-	$("#testmask").Type("abdefcafebcaefbfcabcdef");
+	$("#testmask").Type("abcafebcaefbfcabcdef");
 
 	assert.equal(testmask.value, "abcabcabcabcdef", "Result " + testmask.value);
 });
 
-qunit.test("inputmask(\"Regex\", {regex: \"[а-яА-Я\\s]*\"}) - type space - SilentImp", function(assert) {
+qunit.test("inputmask({regex: \"[а-яА-Я\\s]*\"}) - type space - SilentImp", function(assert) {
 	var $fixture = $("#qunit-fixture");
 	$fixture.append('<input type="text" id="testmask" />');
 	var testmask = document.getElementById("testmask");
-	Inputmask("Regex", {
+	Inputmask({
 		regex: "[а-яА-Я\\s]*"
 	}).mask(testmask);